feat: move ViewRegistry to Prism.Core
