+++ src/Maui/Prism.Maui/Regions/RegionViewRegistry.cs
@@ -3,6 +3,7 @@ using System.Reflection;
 using Prism.Common;
 using Prism.Events;
 using Prism.Ioc;
+using Prism.Mvvm;
 using Prism.Properties;
 
 namespace Prism.Regions;
