+++ src/Prism.Core/Common/IRegistryAware.cs
