+++ src/Prism.Core/Properties/IsExternalInit.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Linq;
+
+#if NETSTANDARD || NET47
+namespace System.Runtime.CompilerServices;
+
+internal static class IsExternalInit
+{
+}
+#endif
